refactor: 변경된 type 및 컴포넌트 사항 반영
